Remove Ledger.getBlocksFrom
The Ledger shouldn't need to worry about older blocks,
as its role is to keep track of the current state,
which doesn't require to know about old blocks once
they have been applied.

As a concession to convenience, we still provide the 'last' block.
An entire unittest section was testing getBlocksFrom and could be removed.
In many other places, the test were redundant, first testing the height,
then getting the blocks and testing the length, and sometimes finally
testing the last block's header.