Validate the random_seed in the Ledger
The random seed validation does not belong in the validation module,
as it requires much more context that we should give to this module.
Instead of getting a value and passing it down a few functions,
just compare it in the caller. Later on we can give better diagnostics
about why the random seed differs.