Test: Use valid blocks in LedgerTest, correct value of ValidatorCycle
This fixes two major issues with the test suite that could not be fixed independently.

First, the ledger tests were using invalid blocks in some places, which wasn't
detected for various reasons, the main one being that
a validator_cycle of 1008 was used in some places,
which itself was inconsistent with the cycle of 20 declared in the test Genesis block.

This inconsistency was also present in other tests (e.g. EnrollmentManager),
and has been fixed. In addition, a check has been applied to the Ledger so that
it cannot creep back.

In the process of fixing this, we now needed to generate valid blocks to feed the Ledger.
This was done by re-working the helper functions used in unittests to add the Enrollments
when needed, as well as to propagate the pre-images.
This approach is still not optimal, as the Ledger always know all pre-images for an enrolled
validator, but it at least allows to move forward and make the tests more correct.

A future enhancement to the code will further simplify the handling of pre-image,
and remove this special-casing, so that tests are more realistic.